move task types to store/types
