+++ src/store/types/commons.ts
@@ -4,3 +4,9 @@ export type BusinessError = {
 };
 
 export type Status = "pending" | "completed" | "error" | "";
+
+export type Pagination = {
+	limit: number;
+	skip: number;
+	total: number;
+};
